pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meccaai"
version = "0.1.0"
description = "AI tool framework with MCP support"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "MeccaAI Team"},
]
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "httpx>=0.24.0",
    "openai>=1.0.0",
    "structlog>=23.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "anyio>=3.6.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "pre-commit>=3.0.0",
]
crewai = [
    "crewai>=0.28.0",
]
langgraph = [
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
]
vector = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]

[project.scripts]
meccaai = "meccaai.apps.openai_app:main"
lumos = "meccaai.apps.lumos_app:main"

[tool.setuptools]
packages = ["meccaai"]

[tool.setuptools.package-dir]
meccaai = "meccaai"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings  
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pyright]
include = ["meccaai"]
exclude = ["**/__pycache__"]
typeCheckingMode = "basic"
pythonVersion = "3.12"

[dependency-groups]
dev = [
    "anyio>=4.10.0",
    "pre-commit>=4.2.0",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
]
